code: unlink temporary files after creation
Previously, prwlock would leave temporary files in $TMPDIR. This commit
fixes this by, after opening a file descriptor to the temporary file,
unlinking the file. This way, as soon as all processes that share the
file descriptor close it, the kernel makes sure to deallocate the space
used by this temporary file. This commit fixes #1.